LPS-86037 Move com.liferay.portal.kernel.util.ObjectGraphUtil to com.liferay.petra.reflect.ObjectGraphUtil
