+++ modules/core/petra/petra-reflect/src/main/java/com/liferay/petra/reflect/ObjectGraphUtil.java
@@ -12,9 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
-
-import com.liferay.petra.reflect.ReflectionUtil;
+package com.liferay.petra.reflect;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Array;
@@ -29,10 +27,7 @@ import java.util.Set;
 
 /**
  * @author Shuyang Zhou
- * @deprecated As of Judson (7.1.x), replaced by {@link
- *             com.liferay.petra.reflect.ObjectGraphUtil}
  */
-@Deprecated
 public class ObjectGraphUtil {
 
 	public static void walkObjectGraph(Object object, Visitor visitor) {
